Integrate style-spec 'auto' property for symbol pitch and rotation
This updates the symbol style layer so that it uses the correct
values (viewport or map) when the auto value is set explicitly
or by default for the properties that support it.

text-pitch-alignment continues to parrot text-rotation-alignment but
now also does so if its value is explicitly set to auto.

This also updates the mapbox-gl-style-spec to use the latest from
the mapbox-gl-style-spec symbol-complex-default branch and includes
a rebuilt v8.min.json.